backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
pydantic
python-dotenv
